+++ Source/BugsnagClient.m
@@ -219,6 +219,11 @@ void BSGWriteSessionCrashData(BugsnagSession *session) {
 @property(readonly, strong, nullable) BugsnagBreadcrumbs *breadcrumbs;
 @end
 
+@interface BugsnagMetadata ()
+- (NSDictionary *_Nonnull)toDictionary;
+@property(unsafe_unretained) id<BugsnagMetadataDelegate> _Nullable delegate;
+@end
+
 @implementation BugsnagClient
 
 @synthesize configuration;