+++ Tests/BugsnagMetadataTests.m
@@ -18,6 +18,8 @@
 
 @interface BugsnagMetadata ()
 @property(atomic, strong) NSMutableDictionary *dictionary;
+- (NSDictionary *_Nonnull)toDictionary;
+@property(unsafe_unretained) id<BugsnagMetadataDelegate> _Nullable delegate;
 @end
 
 // MARK: - DummyClass